feat(autocomplete-multi): keep focus after selecting an option
