+++ src/components/autocomplete/react/Autocomplete.tsx
@@ -8,6 +8,8 @@ import { COMPONENT_PREFIX } from 'LumX/core/react/constants';
 import { handleBasicClasses } from 'LumX/core/utils';
 import { IGenericProps, getRootClassName } from 'LumX/react/utils';
 
+import { useFocusOnClose } from 'LumX/core/react/hooks/useFocusOnClose';
+
 /////////////////////////////
 
 /**
@@ -219,6 +221,7 @@ const Autocomplete: React.FC<AutocompleteProps> = (props: AutocompleteProps): Re
     } = props;
 
     const textFieldRef = useRef(null);
+    useFocusOnClose(inputRef.current, isOpen);
 
     return (
         <div
